Prepare for LSST software installation to be relocatable
Generate shell scripts for setting up the minimal environment required to install LSST software
stack. Bootstrap scripts for the following shells are supported: bash, (t)csh, ksh, zsh. They
are named 'loadLSST.?sh' using an explicit extension to convey their intended use.

The bootstrap scripts allow for the LSST_HOME environment variable to point to a directory
where the LSST software stack is installed. If not already set when the bootstrap script is execute
it sets that variable to the location where the scripts are located and sets up the packages
the stack depends on.

NOTE: this is a combination of several commits to comply with LSST rules for pull requests.