do not build/setup lsst product
The sole function of this product appears to be to setup the
EUPS_PKGROOT env var.  This is no-longer desired as we need to
dynamically generate the EUPS_PKGROOT on platforms for which binary EUPS
"tarballs" are being generated.

Additionally, as a side effect of its use of sconUtils, it requires that
scons/doxygen/etc. be built.  This is unnecessary during a boot strap
phase as any required deps are properly pulled in during a distrib
install.  # Please enter the commit message for your changes. Lines
starting # with '#' will be ignored, and an empty message aborts the
commit.