boostrap miniconda directly
Instead of installing it as an EUPS distrib product.  This removes the
need to have a pre-existing python environment from which to get EUPS
running.  It also ensures that if miniconda is installed, it will be the
interpreter used by EUPS.  # Please enter the commit message for your
changes. Lines starting