Simplify test for existing environment.
The test using `conda info --envs --json` didn't work properly anyway,
as the output included the environment paths, not names (unless the
environment happened to be activated already).

Using `conda env list` output would require a lot of trimming and is slow,
but it might be a bit safer with respect to conda internals changes
(although the command output could change, too, as it's meant for
humans, not machines).  Hope that we don't have to go there.